fix hash diff is always 0

+++ main/thermal/auto_tune.h
@@ -7,7 +7,7 @@ typedef struct
 {
     double power_limit;
     uint16_t fan_limit;
-    uint8_t step_volt;
+    double step_volt;
     double step_freq_rampup;
     double step_freq;
     double autotune_step_frequency;
@@ -15,13 +15,13 @@ typedef struct
     uint16_t max_voltage_asic;
     uint16_t max_frequency_asic;
     uint8_t max_asic_temperatur;
-    uint16_t frequency;
-    uint16_t voltage;
+    double frequency;
+    double voltage;
 } auto_tune_settings;
 
 extern auto_tune_settings AUTO_TUNE;
 void auto_tune_init();
 void auto_tune(bool pid_control_fanspeed);
-uint16_t auto_tune_get_frequency();
-uint16_t auto_tune_get_voltage();
+double auto_tune_get_frequency();
+double auto_tune_get_voltage();
 #endif
